% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_render_data_table.R
\name{mod_render_data_table_ui}
\alias{mod_render_data_table_ui}
\title{render_data_table UI Function}
\usage{
mod_render_data_table_ui(id, type_of_data, title_icon = "magic")
}
\arguments{
\item{id}{A string id that will be used to create a namespace.}

\item{type_of_data}{A string description of the type of data used as the box title.}

\item{title_icon}{String indicating the type of icon to be added to the box title.
Note that only existing fontawesome icons can be used.}
}
\value{
UI table element.
}
\description{
This function has to be set in the UI part of a shiny 
application. It adds a container for a table. The table will be located
inside the shinydashboard::box for which the user can set the title and
the accompanying icon.

A shiny Module that allows users to create a container for table,
and the accompanying server logic to render the table.
}
\examples{
\dontrun{
# In UI :
mod_render_data_table_ui('potions_table', type_of_data = 'potions', title_icon = 'wine-bottle')
# In Server
mod_render_data_table_server(id = 'potions_table', my_data = myGolemApp::potions)
}

}
